fix comparison with unsigned warning

+++ rcl_lifecycle/test/test_default_state_machine.cpp
@@ -90,9 +90,9 @@ TEST_F(TestDefaultStateMachine, zero_init) {
   rcl_lifecycle_state_machine_t state_machine = rcl_lifecycle_get_zero_initialized_state_machine();
   ASSERT_EQ(rcl_lifecycle_state_machine_is_initialized(&state_machine), RCL_RET_ERROR);
   const rcl_lifecycle_transition_map_t * transition_map = &state_machine.transition_map;
-  ASSERT_EQ(transition_map->states_size, 0);
+  ASSERT_EQ(transition_map->states_size, (unsigned int)0);
   ASSERT_EQ(transition_map->states, nullptr);
-  ASSERT_EQ(transition_map->transitions_size, 0);
+  ASSERT_EQ(transition_map->transitions_size, (unsigned int)0);
   ASSERT_EQ(transition_map->transitions, nullptr);
 }
